Implement APIScriptingEngine in API impl
Let platforms be agnostic.
